+++ test/integration/sablier-v2-linear/SablierV2LinearTest.t.sol
@@ -17,8 +17,8 @@ abstract contract SablierV2LinearTest is IntegrationTest {
     //////////////////////////////////////////////////////////////////////////*/
 
     uint40 internal constant TIME_OFFSET = 2_600 seconds;
-    uint256 internal immutable WITHDRAW_AMOUNT_DAI = 2_600e18;
-    uint256 internal immutable WITHDRAW_AMOUNT_USDC = 2_600e6;
+    uint128 internal immutable WITHDRAW_AMOUNT_DAI = 2_600e18;
+    uint128 internal immutable WITHDRAW_AMOUNT_USDC = 2_600e6;
 
     /*//////////////////////////////////////////////////////////////////////////
                                   TESTING VARIABLES
