+++ test/integration/sablier-v2-linear/get-withdrawable-amount/getWithdrawableAmount.t.sol
@@ -17,8 +17,8 @@ contract GetWithdrawableAmount__Test is SablierV2LinearTest {
     /// @dev When the stream does not exist, it should return zero.
     function testGetWithdrawableAmount__StreamNonExistent() external {
         uint256 nonStreamId = 1729;
-        uint256 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(nonStreamId);
-        uint256 expectedWithdrawableAmount = 0;
+        uint128 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(nonStreamId);
+        uint128 expectedWithdrawableAmount = 0;
         assertEq(actualWithdrawableAmount, expectedWithdrawableAmount);
     }
 
@@ -29,16 +29,16 @@ contract GetWithdrawableAmount__Test is SablierV2LinearTest {
     /// @dev it should return zero.
     function testGetWithdrawableAmount__CliffTimeGreaterThanBlockTimestamp() external StreamExistent {
         vm.warp(daiStream.cliffTime - 1 seconds);
-        uint256 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
-        uint256 expectedWithdrawableAmount = 0;
+        uint128 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
+        uint128 expectedWithdrawableAmount = 0;
         assertEq(actualWithdrawableAmount, expectedWithdrawableAmount);
     }
 
     /// @dev When the cliff time is equal to the block timestamp, it should return the correct withdrawable amount.
     function testGetWithdrawableAmount__CliffTimeEqualToBlockTimestamp() external {
         vm.warp(daiStream.cliffTime);
-        uint256 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
-        uint256 expectedWithdrawableAmount = WITHDRAW_AMOUNT_DAI - 100e18;
+        uint128 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
+        uint128 expectedWithdrawableAmount = WITHDRAW_AMOUNT_DAI - 100e18;
         assertEq(actualWithdrawableAmount, expectedWithdrawableAmount);
     }
 
@@ -54,8 +54,8 @@ contract GetWithdrawableAmount__Test is SablierV2LinearTest {
     {
         vm.warp(daiStream.stopTime + 1 seconds);
         sablierV2Linear.withdraw(daiStreamId, WITHDRAW_AMOUNT_DAI);
-        uint256 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
-        uint256 expectedWithdrawableAmount = daiStream.depositAmount - WITHDRAW_AMOUNT_DAI;
+        uint128 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
+        uint128 expectedWithdrawableAmount = daiStream.depositAmount - WITHDRAW_AMOUNT_DAI;
         assertEq(actualWithdrawableAmount, expectedWithdrawableAmount);
     }
 
@@ -66,8 +66,8 @@ contract GetWithdrawableAmount__Test is SablierV2LinearTest {
         CliffTimeLessThanBlockTimestamp
     {
         vm.warp(daiStream.stopTime + 1 seconds);
-        uint256 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
-        uint256 expectedWithdrawableAmount = daiStream.depositAmount;
+        uint128 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
+        uint128 expectedWithdrawableAmount = daiStream.depositAmount;
         assertEq(actualWithdrawableAmount, expectedWithdrawableAmount);
     }
 
@@ -79,8 +79,8 @@ contract GetWithdrawableAmount__Test is SablierV2LinearTest {
     {
         vm.warp(daiStream.stopTime);
         sablierV2Linear.withdraw(daiStreamId, WITHDRAW_AMOUNT_DAI);
-        uint256 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
-        uint256 expectedWithdrawableAmount = daiStream.depositAmount - WITHDRAW_AMOUNT_DAI;
+        uint128 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
+        uint128 expectedWithdrawableAmount = daiStream.depositAmount - WITHDRAW_AMOUNT_DAI;
         assertEq(actualWithdrawableAmount, expectedWithdrawableAmount);
     }
 
@@ -91,8 +91,8 @@ contract GetWithdrawableAmount__Test is SablierV2LinearTest {
         CliffTimeLessThanBlockTimestamp
     {
         vm.warp(daiStream.stopTime);
-        uint256 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
-        uint256 expectedWithdrawableAmount = daiStream.depositAmount;
+        uint128 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
+        uint128 expectedWithdrawableAmount = daiStream.depositAmount;
         assertEq(actualWithdrawableAmount, expectedWithdrawableAmount);
     }
 
@@ -104,8 +104,8 @@ contract GetWithdrawableAmount__Test is SablierV2LinearTest {
     {
         vm.warp(daiStream.startTime + TIME_OFFSET);
         sablierV2Linear.withdraw(daiStreamId, WITHDRAW_AMOUNT_DAI);
-        uint256 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
-        uint256 expectedWithdrawableAmount = 0;
+        uint128 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
+        uint128 expectedWithdrawableAmount = 0;
         assertEq(actualWithdrawableAmount, expectedWithdrawableAmount);
     }
 
@@ -117,8 +117,8 @@ contract GetWithdrawableAmount__Test is SablierV2LinearTest {
     {
         uint256 usdcStreamId = createDefaultUsdcStream();
         vm.warp(usdcStream.startTime + TIME_OFFSET);
-        uint256 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(usdcStreamId);
-        uint256 expectedWithdrawableAmount = WITHDRAW_AMOUNT_USDC;
+        uint128 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(usdcStreamId);
+        uint128 expectedWithdrawableAmount = WITHDRAW_AMOUNT_USDC;
         assertEq(actualWithdrawableAmount, expectedWithdrawableAmount);
     }
 
@@ -129,8 +129,8 @@ contract GetWithdrawableAmount__Test is SablierV2LinearTest {
         CliffTimeLessThanBlockTimestamp
     {
         vm.warp(daiStream.startTime + TIME_OFFSET);
-        uint256 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
-        uint256 expectedWithdrawableAmount = WITHDRAW_AMOUNT_DAI;
+        uint128 actualWithdrawableAmount = sablierV2Linear.getWithdrawableAmount(daiStreamId);
+        uint128 expectedWithdrawableAmount = WITHDRAW_AMOUNT_DAI;
         assertEq(actualWithdrawableAmount, expectedWithdrawableAmount);
     }
 }
